Started working on the Transcribe React App
Modified the TranscribeController so that clicking the
Step 1 link in the Transcribe page links to the new
Transcribe App. This is a temporary solution for testing
purposes. Eventually the Transcribe page will be replaced
with the new React App

Currently the Transcribe app contains only a main App
component and a VideoPlayer component. Both of which
are missing the majority of their functionality