Refactored the implementation of moving the progress bar
Previously, the progress bar would change width as the video
progressed, this isn't ideal as it causes the DOM to recalculate the
positioning of the rest of the page.

Now, the background gradient changes as the video progresses, and
the dimensions of the DOM element stay the same. This means that
only the colours have to repainted and the layout doesn't have to be
recalculated each render. This is much cheaper computationally.