The VideoPlayer component API has been updated to inform it's parent of it's progress
The API for the VideoPlayer now requires a mutable reference to a
number. The VideoPlayer component will keep this reference up to date
with the progress through the video as a fraction.

This is important as now the parent is informed about the Video's state.

Imortantly, it is not a part of the parent's state - the VideoPlayer is
responsible for maintaining it, it simply informs it's parent about it's
state.