Users can now transcribe in parallel
Whenever a change is made to the chunks state provided by
useOurstoryServer the change will be based on what's stored in the
server. In other words, the system uses the api as the source of truth,
rather than it's own internal state.

In future, the chunks state could be replaced completely with calls to
the api, rather than storing the state internally. This would make the
app more live and editing transcriptions in parallel more reliable