Changed API for Header component
Header now accepts a title. This change was made so that the three Steps
could have their own Header title.

Changed the title for the ChunkCard component to be the length and start
of the chunk, rather than the start and end time stamps of the chunks.

Changed the ChunkEditor's list of ChunkCard's to be 2.5 ChunkCards wide,
rather than 3.5. This means that each ChunkCard has more screen
realestate