Added IndabaLink
This is a simple wrapper component around Material UI's Link component.
The rationale for this component is to keep consistent styling of Link
components across the App, as well as to encapsulate the dependency on
Material UI's Link component to only one point.

i.e. if we choose to ditch Material UI we won't have to change all the
Link's in the Transcriber, only IndabaLink's implementation.