Added Completion transaction to Review/Transcribe
The Reviewer and Transcriber components now have a transaction for
completing Reviewing/Transcribing.

This change involved modifying the Slideshow component API - Slideshow
now accepts an optional onComplete function. If it receives this
function it will display a tick arrow instead of a disabled next button
at the end of the slideshow and call this function when the tick arrow
button is clicked.

Also added the BackButton component for consistent back buttons
throughout the app.

The ChunkEditor, Transcriber and Reviewer use the new BackButton
component and all accept atExit instead of a back button component to
handle going back to the Dashboard