Rewrote Slideshow to have a simpler API
Slideshow now no longer supports onNavForward and onNavBack,
instead, it has a simple onNav function.

Added useFirstRender() hook to check if the current render is the first
render. This is currently used in Transcriber to check if the
transcription should be saved, as no transcription should be saved on
the first render