Added Loading Modal and Onboarding Modal
All of the three steps now have a loading modal associated with them so
that the user knows cannot interact with the content of the page until
the video begins buffering.

Implemented an Onboarding Modal for ChunkEditor component, this replaces
the old Information Modal. ChunkEditor uses local storage to remember if
the user has been onboarded.

Eventually, all three of the major steps will have onboarding modals.
This will require the onboarding modal component to be rewritten to be
generic